src/components/layout: switch on media
